fix: security bug patched

+++ src/main/java/com/kenryhraval/banking/dto/TransactionResponse.java
@@ -13,19 +13,19 @@ import java.time.LocalDateTime;
 public class TransactionResponse {
 //    private Long accountId;
     private TransactionType type;
-    private BigDecimal amount;
+    private double amount;
     private LocalDateTime timestamp;
     private String description;
-    private BigDecimal balanceAfter;
+    private double balanceAfter;
     private Long relatedAccountId;
 
     public TransactionResponse(Transaction transaction) {
 //        this.accountId = transaction.getAccount().getId();
         this.type = transaction.getType();
-        this.amount = transaction.getAmount();
+        this.amount = transaction.getAmount().doubleValue();
         this.timestamp = transaction.getTimestamp();
         this.description = transaction.getDescription();
-        this.balanceAfter = transaction.getBalanceAfter();
+        this.balanceAfter = transaction.getBalanceAfter().doubleValue();
         this.relatedAccountId = (transaction.getRelatedAccount() != null) ? transaction.getRelatedAccount().getId() : null;
     }
 }
